Fix beatmap info graph error.

+++ osu.Game.Rulesets.Karaoke/Statistics/BeatmapInfoGraph.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) andy840119 <andy840119@gmail.com>. Licensed under the GPL Licence.
 // See the LICENCE file in the repository root for full licence text.
 
+using System.Collections.Generic;
 using System.Linq;
 using JetBrains.Annotations;
 using osu.Framework.Allocation;
@@ -8,6 +9,7 @@ using osu.Framework.Bindables;
 using osu.Framework.Graphics.Containers;
 using osu.Framework.Input.Events;
 using osu.Game.Beatmaps;
+using osu.Game.Rulesets.Mods;
 using static osu.Game.Screens.Select.BeatmapInfoWedge;
 
 namespace osu.Game.Rulesets.Karaoke.Statistics
@@ -16,6 +18,9 @@ namespace osu.Game.Rulesets.Karaoke.Statistics
     {
         [Resolved(CanBeNull = true)]
         private OsuGame game { get; set; }
+
+        [Resolved]
+        private IBindable<IReadOnlyList<Mod>> mods { get; set; }
 
         private readonly IBeatmap beatmap;
 
@@ -39,13 +44,13 @@ namespace osu.Game.Rulesets.Karaoke.Statistics
                 return;
 
             var beatmapDifficulty = difficultyCache.GetBindableDifficulty(beatmap.BeatmapInfo).Value;
-            LoadComponentAsync(new BeatmapInfoWedge(workingBeatmap.Value, beatmapDifficulty), Add);
+            LoadComponentAsync(new BeatmapInfoWedge(workingBeatmap.Value, mods.Value, beatmapDifficulty ?? new StarDifficulty()), Add);
         }
 
         public class BeatmapInfoWedge : BufferedWedgeInfo
         {
-            public BeatmapInfoWedge(WorkingBeatmap beatmap, StarDifficulty difficulty)
-                : base(beatmap, new KaraokeRuleset().RulesetInfo, difficulty)
+            public BeatmapInfoWedge(WorkingBeatmap beatmap, IReadOnlyList<Mod> mods, StarDifficulty difficulty)
+                : base(beatmap, new KaraokeRuleset().RulesetInfo, mods, difficulty)
             {
             }
 
@@ -57,10 +62,17 @@ namespace osu.Game.Rulesets.Karaoke.Statistics
                 if (centerMetadata != null)
                     centerMetadata.Y = -20;
 
+                // Use tricky to add extra info.
+                if (InfoLabelContainer == null)
+                    return;
+
                 var shouldBeRemovedLabel = InfoLabelContainer.Children.OfType<InfoLabel>()
                                                              .Where(x => x.TooltipText == "Note" || x.TooltipText == "This beatmap is not scorable.").ToList();
                 InfoLabelContainer.RemoveRange(shouldBeRemovedLabel);
             }
+
+            protected FillFlowContainer InfoLabelContainer
+                => (Children.LastOrDefault() as FillFlowContainer)?.LastOrDefault() as FillFlowContainer;
         }
     }
 }
